+++ integration-tests/general/common/src/main/kotlin/com/r3/conclave/integrationtests/general/common/EnclaveContext.kt
@@ -26,4 +26,5 @@ abstract class EnclaveContext {
     abstract fun getSecretKey(keyRequest: Cursor<SgxKeyRequest, ByteBuffer>): ByteArray
     abstract fun sealData(data: PlaintextAndAD): ByteArray
     abstract fun unsealData(sealedBlob: ByteArray): PlaintextAndAD
+    abstract fun createAttestationQuote(reportData: ByteArray): ByteArray
 }